backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
lightgbm==4.1.0
shap==0.43.0
matplotlib==3.8.2
scikit-learn==1.3.2
anthropic==0.7.8
python-dotenv==1.0.0
binance-futures-connector==3.3.1
tqdm==4.66.1
ta-lib==0.4.28
